api/requirements.txt
fastapi>=0.112.4
uvicorn>=0.25.0
python-dotenv>=1.0.1
azure-cosmos>=4.7.0
azure-identity>=1.19.0
azure-communication-email==1.0.0
colorlog>=6.8.2
requests>=2.32.3
invoke>=2.2.0
starlette_gzip_request>=0.1.0
azure-ai-projects>=1.0.0b12
openai>=1.0.0